Crate gooey_browser

A Frontend for Gooey that targets web browsers by creating DOM elements using web-sys and wasm-bindgen.

Window implementation

The Window implementation in this frontend is limited by the browser’s APIs. Of note:

maximized() and set_maximized() interact with the fullscreen APIs.

inner_position() and outer_position() are equivalent. This also means that set_outer_position() actually sets the inner position.

Controlling the browser window (closing, resizing, moving) is dependent upon the browser allowing the operation. This generally is only possible if the window is opened via a javascript API and has no other tabs within in.

Modules

utils

Macros

make_browser

Structs

RegisteredTransmogrifier

WebSys

WebSys is a Frontend implementor that implements Gooey interfaces through web-sys in a web browser.

WidgetClosure

Traits

AnyWebSysTransmogrifier

ImageExt

WebSysTransmogrifier