Struct gooey_browser::WebSys

pub struct WebSys {
    pub ui: Gooey<Self>,
    // some fields omitted
}

WebSys is a Frontend implementor that implements Gooey interfaces through web-sys in a web browser.

Fields
ui: Gooey<Self>
Implementations

impl WebSys

pub fn initialize()

#[must_use]
pub fn new(ui: Gooey<Self>, configuration: Configuration) -> Self

pub fn install_in_id(&mut self, id: &str)

pub fn with_transmogrifier<TResult, C: FnOnce(&dyn AnyWebSysTransmogrifier, AnyTransmogrifierContext<'_, Self>) -> TResult>(
&self,
widget_id: &WidgetId,
callback: C
) -> Option<TResult>

Executes callback with the transmogrifier and transmogrifier state as parameters.

pub fn register_intent_handler(
&self,
intent: Intent,
widget: &WidgetRegistration,
callback: Callback
)